Move function to new folder
